+++ packages/odata-service-inquirer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/odata-service-inquirer
 
+## 0.5.42
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.27.4
+-   @sap-ux/telemetry@0.5.30
+-   @sap-ux/axios-extension@1.16.6
+
 ## 0.5.41
 
 ### Patch Changes